+++ contracts/modules/InscriptionModule.sol
@@ -21,10 +21,10 @@ contract InscriptionModule is IInscriptionModule {
      * @notice Inscribe a message onto this bot.
      * The message is in the calldata.
      */
-    fallback () external payable override {}
+    fallback () external payable virtual override {}
 
     /**
      * @notice Allows this contract to receive the gas token.
      */
-    receive () external payable override {}
+    receive () external payable virtual override {}
 }
